requirements.txt
asgiref==3.8.1
Django==5.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.9.0
python-decouple==3.8